displays the message only once complete string is recieved

+++ src/client.c
@@ -6,7 +6,7 @@
 /*   By: taospa <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/07 21:39:59 by taospa            #+#    #+#             */
-/*   Updated: 2023/08/09 15:17:05 by tsaint-p         ###   ########.fr       */
+/*   Updated: 2023/08/10 15:42:25 by tsaint-p         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,37 +14,72 @@
 #include <signal.h>
 #include "../libft/libft.h"
 
+int	send_nul_byte(pid_t pid)
+{
+	int	success;
+	int	i;
+
+	success = 0;
+	i = 0;
+	while (i < 8)
+	{
+		success = kill(pid, SIGUSR1);
+		if (success == -1)
+			return (-1);
+		i++;
+		usleep(400);
+	}
+	return (0);
+}
+
+int	send_len(char *str, pid_t pid)
+{
+	unsigned int	ref;
+	size_t			len;
+	int				success;
+
+	success = 0;
+	len = ft_strlen(str);
+	ref = 0b10000000000000000000000000000000;
+	while (ref)
+	{
+		if (!(len & ref))
+			success = kill(pid, SIGUSR1);
+		else
+			success = kill(pid, SIGUSR2);
+		ref = ref >> 1;
+		usleep(400);
+	}
+	return (success);
+}
+
 int	send_sig(char *str, pid_t pid)
 {
-	int	i;
 	int	ref;
+	int	success;
 
-	i = 0;
+	success = 0;
 	ref = 0b10000000;
+	if (send_len(str, pid) == -1)
+		return (-1);
 	while (*str)
 	{
-		while (i < 8)
+		while (ref)
 		{
 			if (!(*str & ref))
-			{
-				//write(1, "0", 1);
-				kill(pid, SIGUSR1);
-			}
+				success = kill(pid, SIGUSR1);
 			else
-			{
-				kill(pid, SIGUSR2);
-				//write(1, "1", 1);
-			}
-			i++;
+				success = kill(pid, SIGUSR2);
 			ref = ref >> 1;
-			usleep(500);
+			if (success == -1)
+				return (-1);
+			usleep(400);
 		}
-		//write(1, "\n", 1);
 		ref = 0b10000000;
-		i = 0;
 		str++;
 	}
-	return (1);
+	success = send_nul_byte(pid);
+	return (success);
 }
 
 int	main(int argc, char **argv)
@@ -54,7 +89,7 @@ int	main(int argc, char **argv)
 	if (argc != 3)
 		return (write(2, "Invalid argments\n", 17));
 	pid = ft_atoi(argv[1]);
-	if (send_sig(argv[2], pid))
+	if (pid && send_sig(argv[2], pid) != -1)
 		write(1, "Signal sent\n", 12);
 	else
 		write(2, "Error sending signal, make sure you used a valid PID\n", 53);
